[Decomp][Player] Cohesion split: extract gear-score calculation to PlayerGearScore.cpp
